[AUDIT-3] third-pass PostgreSQL review — 14 more bugs fixed across 12 files
contacts.ts (CRITICAL): safetyOfficerType was destructured from req.body but
  absent from params array, SET clause, and all 5 SELECT queries. Every admin
  edit silently discarded the Officer Type value; the dropdown always showed
  blank on reopen. Fixed in PUT (added $22), all 5 SELECTs, and ContactRow
  interface.

auth.ts (HIGH — 3 bugs):
  - DELETE /users/:id had TOCTOU race: two concurrent requests against the two
    remaining admins could both pass the count check and both succeed, leaving
    zero admins. Wrapped SELECT+UPDATE+DELETE in transaction() with FOR UPDATE.
  - sendPasswordResetEmail: DELETE old tokens + INSERT new token were separate
    execute() calls. Crash between them = user has no valid reset token. Wrapped
    in transaction().
  - sendPasswordResetEmail: contact.name/surname interpolated into HTML email
    without escaping — XSS via contact name. Added escapeHtml() helper.

sites/crud.ts + closures.ts (HIGH): Closure date queries used UTC 'today' via
  new Date().toISOString().split('T')[0]. Between midnight and ~10-11am AEST
  this returned yesterday — closures not shown, stale day visible. All 4
  occurrences changed to toLocaleDateString('en-CA', {timeZone: 'Australia/Melbourne'}).

sites/helpers.ts: SITES_COLUMNS was still missing 3 real DB columns added in
  migration 009 (heroImages, displayOnMap, displayInList). Pass-1 fixed ghost
  columns and 3 missing ones; these 3 were still absent. Count now 54=54.

realMessageService.ts (HIGH): thumbsUp, thumbsDown, markDelivered all used
  WHERE id = $1 with string params against SERIAL column — crashed with
  'operator does not exist: integer = text' on every reaction. Added ::int cast
  at all 6 query sites.

realRetrievalService.ts: requestRetrieval and createRetrievalForPilot did
  SELECT-then-INSERT outside a transaction — concurrent requests could both
  pass the existence check and create duplicate retrievals. Wrapped in
  transaction().

publicContacts.ts: Pilot DELETE ran 5 sequential execute() calls (reset tokens,
  sessions, breadcrumbs, flights, pilot) with no transaction — partial deletion
  left inconsistent state. Wrapped in transaction().

projects.ts + pages.ts: DELETE sequences (project contacts+documents+project,
  page+attachments) had no transaction. Wrapped both.

pilotAuth.ts: optionalPilotAuth didn't delete expired sessions; requirePilotAuth
  and /me both did. Expired rows accumulated indefinitely. Added cleanup.

demoRetrievalService.ts: lat || null / lon || null clobbered coordinate 0.
  Changed to ?? null.

+++ server/services/realMessageService.ts
@@ -87,35 +87,35 @@ export class RealMessageService implements MessageService {
 
   async thumbsUp(msgId: string | number, pilotId: string) {
     const msg = await queryOne<{ id: number; recipientPilotId: string }>(
-      `SELECT id, "recipientPilotId" FROM map_messages WHERE id = $1`,
+      `SELECT id, "recipientPilotId" FROM map_messages WHERE id = $1::int`,
       [msgId]
     );
 
     if (!msg) return { ok: false, error: "Message not found", status: 404 };
     if (msg.recipientPilotId !== pilotId) return { ok: false, error: "Not your message", status: 403 };
 
-    await execute(`UPDATE map_messages SET "thumbsUp" = 1 WHERE id = $1`, [msgId]);
+    await execute(`UPDATE map_messages SET "thumbsUp" = 1 WHERE id = $1::int`, [msgId]);
     log.info(`Thumbs-up on message ${msgId}`);
     return { ok: true };
   }
 
   async thumbsDown(msgId: string | number, pilotId: string) {
     const msg = await queryOne<{ id: number; recipientPilotId: string }>(
-      `SELECT id, "recipientPilotId" FROM map_messages WHERE id = $1`,
+      `SELECT id, "recipientPilotId" FROM map_messages WHERE id = $1::int`,
       [msgId]
     );
 
     if (!msg) return { ok: false, error: "Message not found", status: 404 };
     if (msg.recipientPilotId !== pilotId) return { ok: false, error: "Not your message", status: 403 };
 
-    await execute(`UPDATE map_messages SET "thumbsUp" = 2 WHERE id = $1`, [msgId]);
+    await execute(`UPDATE map_messages SET "thumbsUp" = 2 WHERE id = $1::int`, [msgId]);
     log.info(`Thumbs-down on message ${msgId}`);
     return { ok: true };
   }
 
   async markDelivered(msgId: string | number, pilotId: string) {
     const msg = await queryOne<{ id: number; recipientPilotId: string; senderPilotId: string; thumbsUp: number | null }>(
-      `SELECT id, "recipientPilotId", "senderPilotId", "thumbsUp" FROM map_messages WHERE id = $1`,
+      `SELECT id, "recipientPilotId", "senderPilotId", "thumbsUp" FROM map_messages WHERE id = $1::int`,
       [msgId]
     );
 
@@ -128,7 +128,7 @@ export class RealMessageService implements MessageService {
       return { ok: false, error: "Not your message", status: 403 };
     }
 
-    await execute(`UPDATE map_messages SET "deliveredAt" = CURRENT_TIMESTAMP WHERE id = $1`, [msgId]);
+    await execute(`UPDATE map_messages SET "deliveredAt" = CURRENT_TIMESTAMP WHERE id = $1::int`, [msgId]);
     return { ok: true };
   }
 }
